+++ libraries/RS485/RS485.h
@@ -3,7 +3,7 @@
 //    FILE: RS485.h
 //  AUTHOR: Rob Tillaart
 //    DATE: 30-okt-2017
-// VERSION: 0.2.6
+// VERSION: 0.3.0
 // PURPOSE: Arduino library for RS485 modules
 //     URL: https://github.com/RobTillaart/RS485
 
@@ -11,23 +11,13 @@
 #include "Arduino.h"
 #include "ASCII_CONTROL.h"
 
-#define RS485_LIB_VERSION        (F("0.2.6"))
+#define RS485_LIB_VERSION        (F("0.3.0"))
 
 
 class RS485 : public Stream
 {
 public:
   RS485(Stream * stream, uint8_t sendPin, uint8_t deviceID = 0);
-
-  //  sets the baud rate of the stream.
-  //  0.3.0
-  //  void begin(uint32_t baudRate);
-
-  //  baudRate must match Serial baudRate
-  //  micros are needed before RS485 stream may be reset
-  //  to (default) receiving mode.
-  void     setMicrosPerByte(uint32_t baudRate);
-  uint32_t getMicrosPerByte();
 
   uint8_t  getDeviceID();
 
@@ -66,7 +56,7 @@ private:
   uint8_t  _deviceID      = 0;
   uint16_t _microsPerByte = 1000;
 
-  //       EXPERIMENTAL (for 0.3.0)
+  //       EXPERIMENTAL (for 0.3.x)
   uint8_t _bidx = 0;
   uint8_t _buffer[50];  //  internal receive buffer
 };