+++ libraries/RS485/test/unit_test_001.cpp
@@ -62,25 +62,6 @@ unittest(test_constructor)
 }
 
 
-unittest(test_microsPerByte)
-{
-  Serial.begin(115200);
-
-  const uint8_t sendPin  = 4;
-
-  RS485 master(&Serial, sendPin);
-  assertEqual(0, master.getDeviceID());
-
-  //  default value
-  assertEqual(1000, master.getMicrosPerByte());
-
-  master.setMicrosPerByte(115200);
-  //  count 11 bits / byte
-  uint32_t us = (11 * 1000000UL) / 115200;
-  assertEqual(us, master.getMicrosPerByte());
-}
-
-
 unittest(test_communication_mode)
 {
   Serial.begin(115200);